refactor: rename networking -> network
