+++ src/client/controllers/effect-controller.ts
@@ -1,6 +1,6 @@
 import { Controller, OnStart } from "@flamework/core";
 import { EFFECTS } from "client/effects";
-import { Events } from "client/networking";
+import { Events } from "client/network";
 
 @Controller()
 export class EffectController implements OnStart {
