+++ src/client/network/index.ts
@@ -1,4 +1,4 @@
-import { GlobalEvents, GlobalFunctions } from "shared/networking";
+import { GlobalEvents, GlobalFunctions } from "shared/network";
 
 export const Events = GlobalEvents.createClient({});
 export const Functions = GlobalFunctions.createClient({});
