+++ src/shared/network/character.ts
