+++ src/shared/network/dialogue.ts
