+++ src/shared/network/mana.ts
